backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pyswisseph==2.10.3.2
geopy==2.4.1
pytz==2024.1
timezonefinder==5.2.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
python-multipart==0.0.9
